Fix piece moving smoothly
Modified the movePieceSmoothly function to take a PieceLogic instead of a gameObject, so that it is reusable and able to be used from the PieceLogic script, in which we can now use it to put the piece down. Still have to implement the delay and implement the function for other.